+++ src/app/sign-up/page.const.ts
@@ -1,0 +1,4 @@
+export const STEP = {
+  SIGN_UP: 1,
+  EMAIL_VERIFICATION: 2,
+}
